To hell with packaging JS, just dump in a script tag into the main page
I got packaging working somewhat with importmap and esbuild (in another project) after a lot of wasted time
but for whatever reason it's not loaded in time for the page to use it? Timebox over, fuck this
dump a script tag into the page. I only need one package anyway