Scale the image correctly. This approach is not ideal
First, I think manually specifying the dimensions is bugged. So changed to use a preset
The canvas is crap at scaling images. If we load the full thing and let the canvas scale it
the result looks poor.
If we scale the image down to fit the canvas the result looks low resolution